refactor: component명 pascal case 로 변경
