+++ src/pages/NewProject/NewProjectContainer.tsx
@@ -1,9 +1,8 @@
-import React , { useState } from 'react';
+import React, { useState } from 'react';
 import { User } from '@/state/type';
-import NewProjectForm from './newProjectForm';
+import NewProjectForm from './NewProjectForm';
 
-
-const NewProjectContainer = () : JSX.Element => {
+const NewProjectContainer = (): JSX.Element => {
   const [seletedPlatform, setPlatform] = useState('아직 선택하지 않았습니다.');
   const [projectName, setProjectName] = useState('New Project');
   const [alertSetting, setAlert] = useState('거부');
@@ -12,15 +11,19 @@ const NewProjectContainer = () : JSX.Element => {
   const [projectAdmins, setAdmins] = useState<User[]>([]);
 
   const prop = {
-    seletedPlatform,setPlatform,
-    projectName,setProjectName,
-    alertSetting,setAlert,
-    alertMails,setMails,
-    projectMembers,setMembers,
-    projectAdmins, setAdmins
+    seletedPlatform,
+    setPlatform,
+    projectName,
+    setProjectName,
+    alertSetting,
+    setAlert,
+    alertMails,
+    setMails,
+    projectMembers,
+    setMembers,
+    projectAdmins,
+    setAdmins,
   };
-  return (
-    <NewProjectForm {...prop} />
-  );
-}
+  return <NewProjectForm {...prop} />;
+};
 export default NewProjectContainer;
