+++ src/pages/NewProject/components/PlatformSelector.tsx
